[ghstack-poisoned]

+++ kernels/portable/cpu/op_repeat.cpp
@@ -6,6 +6,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+#include <c10/util/irange.h>
 #include <cstring>
 
 #include <executorch/kernels/portable/cpu/util/repeat_util.h>
@@ -29,11 +30,11 @@ bool calculate_output_size(
       repeats.size(),
       self_sizes.size());
 
-  int32_t i = 0;
+  size_t i = 0;
   for (; i < (repeats.size() - self_sizes.size()); ++i) {
     out_sizes_ptr[i] = static_cast<executorch::aten::SizesType>(repeats[i]);
   }
-  int32_t j = 0;
+  size_t j = 0;
   for (; i < repeats.size(); ++i) {
     out_sizes_ptr[i] =
         static_cast<executorch::aten::SizesType>(repeats[i]) * self_sizes[j];
